+++ src/OpenTelemetry.Instrumentation.ConfluentKafka/TracerProviderBuilderExtensions.Consumer.cs
@@ -3,7 +3,6 @@
 
 using Confluent.Kafka;
 using Microsoft.Extensions.DependencyInjection;
-using Microsoft.Extensions.Options;
 using OpenTelemetry.Instrumentation.ConfluentKafka;
 using OpenTelemetry.Internal;
 
@@ -31,6 +30,18 @@ public static partial class TracerProviderBuilderExtensions
     /// <typeparam name="TKey">The type of the key.</typeparam>
     /// <typeparam name="TValue">The type of the value.</typeparam>
     /// <param name="builder"><see cref="TracerProviderBuilder"/> being configured.</param>
+    /// <param name="name">The name of the instrumentation.</param>
+    /// <returns>The instance of <see cref="TracerProviderBuilder"/> to chain the calls.</returns>
+    public static TracerProviderBuilder AddKafkaConsumerInstrumentation<TKey, TValue>(
+        this TracerProviderBuilder builder, string? name)
+        => AddKafkaConsumerInstrumentation<TKey, TValue>(builder, name: name, consumerBuilder: null);
+
+    /// <summary>
+    /// Enables automatic data collection of outgoing requests to Kafka.
+    /// </summary>
+    /// <typeparam name="TKey">The type of the key.</typeparam>
+    /// <typeparam name="TValue">The type of the value.</typeparam>
+    /// <param name="builder"><see cref="TracerProviderBuilder"/> being configured.</param>
     /// <param name="consumerBuilder"><see cref="InstrumentedConsumerBuilder{TKey,TValue}"/> to instrument.</param>
     /// <returns>The instance of <see cref="TracerProviderBuilder"/> to chain the calls.</returns>
     public static TracerProviderBuilder AddKafkaConsumerInstrumentation<TKey, TValue>(
@@ -48,7 +59,7 @@ public static partial class TracerProviderBuilderExtensions
     /// <typeparam name="TKey">The type of the key.</typeparam>
     /// <typeparam name="TValue">The type of the value.</typeparam>
     /// <param name="builder"><see cref="TracerProviderBuilder"/> being configured.</param>
-    /// <param name="name">Optional name which is used when retrieving options.</param>
+    /// <param name="name">The name of the instrumentation.</param>
     /// <param name="consumerBuilder">Optional <see cref="InstrumentedConsumerBuilder{TKey, TValue}"/> to instrument.</param>
     /// <returns>The instance of <see cref="TracerProviderBuilder"/> to chain the calls.</returns>
     public static TracerProviderBuilder AddKafkaConsumerInstrumentation<TKey, TValue>(
@@ -58,34 +69,21 @@ public static partial class TracerProviderBuilderExtensions
     {
         Guard.ThrowIfNull(builder);
 
-        name ??= Options.DefaultName;
-
-        builder.ConfigureServices(services =>
-        {
-            services.Configure<ConfluentKafkaConsumerInstrumentationOptions<TKey, TValue>>(name, EnableTracing);
-        });
-
         return builder
             .AddSource(ConfluentKafkaCommon.InstrumentationName)
             .AddInstrumentation(sp =>
             {
-                if (consumerBuilder == null)
+                if (name == null)
                 {
-                    consumerBuilder = sp.GetRequiredService<InstrumentedConsumerBuilder<TKey, TValue>>();
-                    var options = sp.GetRequiredService<IOptionsMonitor<ConfluentKafkaConsumerInstrumentationOptions<TKey, TValue>>>();
-                    consumerBuilder.Options = options.Get(name);
+                    consumerBuilder ??= sp.GetRequiredService<InstrumentedConsumerBuilder<TKey, TValue>>();
+                }
+                else
+                {
+                    consumerBuilder ??= sp.GetRequiredKeyedService<InstrumentedConsumerBuilder<TKey, TValue>>(name);
                 }
 
-                if (consumerBuilder.Options == null)
-                {
-                    consumerBuilder.Options = new ConfluentKafkaConsumerInstrumentationOptions<TKey, TValue>();
-                    EnableTracing(consumerBuilder.Options);
-                }
-
+                consumerBuilder.EnableTraces = true;
                 return new ConfluentKafkaConsumerInstrumentation<TKey, TValue>(consumerBuilder);
             });
     }
-
-    private static void EnableTracing<TKey, TValue>(ConfluentKafkaConsumerInstrumentationOptions<TKey, TValue> options) =>
-        options.Traces = true;
 }