+++ src/OpenTelemetry.Instrumentation.ConfluentKafka/CHANGELOG.md
@@ -2,8 +2,11 @@
 
 ## Unreleased
 
+- Add named instrumentation support
+  ([#2074](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/2074))
+
 ## 0.1.0-alpha.1
 
 Released 2024-Sep-16
 
-* Initial release
+- Initial release